Code quality fixes Added using around every statement for EFU Added NoOptimization for JIT
