+++ EntityFramework.Utilities/PerformanceTests/Models/Address.cs
@@ -1,0 +1,11 @@
+﻿namespace PerformanceTests.Models
+{
+	public class Address
+	{
+		public string Line1 { get; set; }
+
+		public string ZipCode { get; set; }
+
+		public string Town { get; set; }
+	}
+}
