+++ EntityFramework.Utilities/PerformanceTests/Models/Publication.cs
@@ -1,0 +1,13 @@
+﻿namespace PerformanceTests.Models
+{
+	using System.Collections.Generic;
+
+	public class Publication
+	{
+		public int Id { get; set; }
+
+		public string Title { get; set; }
+
+		public ICollection<Comment> Comments { get; set; }
+	}
+}
