+++ EntityFramework.Utilities/PerformanceTests/PerformanceTests.csproj
@@ -47,6 +47,10 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Context.cs" />
+    <Compile Include="Models\Address.cs" />
+    <Compile Include="Models\Comment.cs" />
+    <Compile Include="Models\Publication.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
