+++ EntityFramework.Utilities/PerformanceTests/Program.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using System.Data.Entity;
 using System.Diagnostics;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using EntityFramework.Utilities;
+using PerformanceTests.Models;
 
 namespace PerformanceTests
 {
@@ -25,18 +27,19 @@ namespace PerformanceTests
 			NormalIteration(100000);
 		}
 
+		[MethodImpl(MethodImplOptions.NoOptimization)]
 		private static void NormalIteration(int count)
 		{
 			Console.WriteLine("Standard iteration with " + count + " entities");
 			CreateAndWarmUp();
-			var stop = new Stopwatch();
+			var stopwatch = new Stopwatch();
 
 			using (var db = new Context())
 			{
 				db.Configuration.AutoDetectChangesEnabled = false;
 				db.Configuration.ValidateOnSaveEnabled = false;
 				var comments = GetEntities(count).ToList();
-				stop.Start();
+				stopwatch.Start();
 
 				foreach (var comment in comments)
 				{
@@ -44,23 +47,22 @@ namespace PerformanceTests
 				}
 
 				db.SaveChanges();
-				stop.Stop();
-				Console.WriteLine("Insert entities: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("Insert entities: " + stopwatch.ElapsedMilliseconds + "ms");
 			}
 
 			using (var db = new Context())
 			{
 				db.Configuration.AutoDetectChangesEnabled = true;
 				db.Configuration.ValidateOnSaveEnabled = false;
-				stop.Restart();
-				var toUpdate = db.Comments.Where(c => c.Text == "a").ToList();
-				foreach (var item in toUpdate)
+				stopwatch.Restart();
+				foreach (var item in db.Comments.Where(c => c.Text == "a").ToList())
 				{
 					item.Reads++;
 				}
 				db.SaveChanges();
-				stop.Stop();
-				Console.WriteLine("Update all entities with a: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("Update all entities with a: " + stopwatch.ElapsedMilliseconds + "ms");
 			}
 
 			using (var db = new Context())
@@ -73,83 +75,93 @@ namespace PerformanceTests
 				{
 					item.Reads = rand.Next(0, 9999999);
 				}
-				stop.Restart();
+				stopwatch.Restart();
 				db.SaveChanges();
 
-				stop.Stop();
-				Console.WriteLine("Update all with a random read: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("Update all with a random read: " + stopwatch.ElapsedMilliseconds + "ms");
 			}
 
 			using (var db = new Context())
 			{
 				db.Configuration.AutoDetectChangesEnabled = false;
 				db.Configuration.ValidateOnSaveEnabled = false;
-				stop.Restart();
-				var toDelete = db.Comments.Where(c => c.Text == "a").ToList();
-				foreach (var item in toDelete)
+				stopwatch.Restart();
+				foreach (var item in db.Comments.Where(c => c.Text == "a").ToList())
 				{
 					db.Comments.Remove(item);
 				}
 				db.SaveChanges();
-				stop.Stop();
-				Console.WriteLine("delete all entities with a: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("delete all entities with a: " + stopwatch.ElapsedMilliseconds + "ms");
 			}
 
 			using (var db = new Context())
 			{
 				db.Configuration.AutoDetectChangesEnabled = false;
 				db.Configuration.ValidateOnSaveEnabled = false;
-				stop.Restart();
-				var all = db.Comments.ToList();
-				foreach (var item in all)
+				stopwatch.Restart();
+				foreach (var item in db.Comments.ToList())
 				{
 					db.Comments.Remove(item);
 				}
 				db.SaveChanges();
-				stop.Stop();
-				Console.WriteLine("delete all entities: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("delete all entities: " + stopwatch.ElapsedMilliseconds + "ms");
 			}
 		}
 
+		[MethodImpl(MethodImplOptions.NoOptimization)]
 		private static void BatchIteration(int count)
 		{
 			Console.WriteLine("Batch iteration with " + count + " entities");
 			CreateAndWarmUp();
+			var stopwatch = new Stopwatch();
+			using (var db = new Context())
+			{
+				var comments = GetEntities(count).ToList();
+				stopwatch.Start();
+				EFBatchOperation.For(db, db.Comments).InsertAll(comments);
+				stopwatch.Stop();
+				Console.WriteLine("Insert entities: " + stopwatch.ElapsedMilliseconds + "ms");
+			}
 
 			using (var db = new Context())
 			{
-				var stop = new Stopwatch();
-				var comments = GetEntities(count).ToList();
-				stop.Start();
-				EFBatchOperation.For(db, db.Comments).InsertAll(comments);
-				stop.Stop();
-				Console.WriteLine("Insert entities: " + stop.ElapsedMilliseconds + "ms");
-
-				stop.Restart();
+				stopwatch.Restart();
 				EFBatchOperation.For(db, db.Comments).Where(x => x.Text == "a").Update(x => x.Reads, x => x.Reads + 1);
-				stop.Stop();
-				Console.WriteLine("Update all entities with a: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("Update all entities with a: " + stopwatch.ElapsedMilliseconds + "ms");
+			}
 
+			using (var db = new Context())
+			{
 				var commentsFromDb = db.Comments.AsNoTracking().ToList();
 				var rand = new Random();
 				foreach (var item in commentsFromDb)
 				{
 					item.Reads = rand.Next(0, 9999999);
 				}
-				stop.Restart();
+				stopwatch.Restart();
 				EFBatchOperation.For(db, db.Comments).UpdateAll(commentsFromDb, x => x.ColumnsToUpdate(c => c.Reads));
-				stop.Stop();
-				Console.WriteLine("Bulk update all with a random read: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("Bulk update all with a random read: " + stopwatch.ElapsedMilliseconds + "ms");
+			}
 
-				stop.Restart();
+			using (var db = new Context())
+			{
+				stopwatch.Restart();
 				EFBatchOperation.For(db, db.Comments).Where(x => x.Text == "a").Delete();
-				stop.Stop();
-				Console.WriteLine("delete all entities with a: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("delete all entities with a: " + stopwatch.ElapsedMilliseconds + "ms");
+			}
 
-				stop.Restart();
+			using (var db = new Context())
+			{
+				stopwatch.Restart();
 				EFBatchOperation.For(db, db.Comments).Where(x => true).Delete();
-				stop.Stop();
-				Console.WriteLine("delete all entities: " + stop.ElapsedMilliseconds + "ms");
+				stopwatch.Stop();
+				Console.WriteLine("delete all entities: " + stopwatch.ElapsedMilliseconds + "ms");
 			}
 		}
 
@@ -173,7 +185,7 @@ namespace PerformanceTests
 
 		private static IEnumerable<Comment> GetEntities(int count)
 		{
-			var comments = Enumerable.Repeat('a', count).Select((c, i) => new Comment
+			return Enumerable.Repeat('a', count).Select((c, i) => new Comment
 			{
 				Text = ((char)(c + i % 25)).ToString(),
 				Date = DateTime.Now.AddDays(i),
@@ -184,48 +196,6 @@ namespace PerformanceTests
 					Town = "Town"
 				}
 			});
-			return comments;
 		}
-	}
-
-	public class Context : DbContext
-	{
-		public Context()
-			: base("Data Source=./; Initial Catalog=EFUTest; Integrated Security=SSPI; MultipleActiveResultSets=True") { }
-
-		public IDbSet<Comment> Comments { get; set; }
-		public DbSet<Publication> Publications { get; set; }
-
-		protected override void OnModelCreating(DbModelBuilder modelBuilder)
-		{
-			base.OnModelCreating(modelBuilder);
-
-			modelBuilder.ComplexType<Address>();
-		}
-	}
-
-	public class Comment
-	{
-		public int Id { get; set; }
-		public string Text { get; set; }
-		public DateTime Date { get; set; }
-		public int Reads { get; set; }
-		public Address Address { get; set; }
-		public int? PublicationId { get; set; }
-		public Publication Publication { get; set; }
-	}
-
-	public class Publication
-	{
-		public int Id { get; set; }
-		public string Title { get; set; }
-		public ICollection<Comment> Comments { get; set; }
-	}
-
-	public class Address
-	{
-		public string Line1 { get; set; }
-		public string ZipCode { get; set; }
-		public string Town { get; set; }
 	}
 }
